Crate array_map

Modules

ext

This module contains extension traits that are not yet present in the standard libary.

iter

Macros

array_map

index_map

Structs

ArrayMapFacade

CapacityError

OccupiedEntry

A view into an occupied entry in an ArrayMap. It is part of the Entry enum.

RescaleError

VacantEntry

A view into a vacant entry in an ArrayMap. It is part of the Entry enum.

Enums

Entry

A view into a single entry in a map, which may either be vacant or occupied.

UnavailableMutError

The error type for ArrayMap::get_each_value_mut and ArrayMap::get_each_key_value_mut.

Traits

ArrayMapExt

Type Definitions

ArrayMap

DefaultHashBuilder

Default hasher for ArrayMapFacade.

IndexMap