Module array_map::map

Modules

iter

Structs

ArrayMapFacade

IndexOutOfBoundsError

OccupiedEntry

A view into an occupied entry in an ArrayMap. It is part of the Entry enum.

VacantEntry

A view into a vacant entry in an ArrayMap. It is part of the Entry enum.

Enums

Entry

A view into a single entry in a map, which may either be vacant or occupied.

Type Definitions

ArrayMap

DefaultHashBuilder

Default hasher for ArrayMapFacade.

IndexMap