Struct array_map::set::ArraySetFacade

pub struct ArraySetFacade<T, R: RawTable<(T, ())>, B = DefaultHashBuilder> { /* fields omitted */ }

Implementations

impl<T, R: RawTable<(T, ())> + Default> ArraySetFacade<T, R, DefaultHashBuilder>

pub fn new() -> Self

Creates an empty ArraySetFacade with the DefaultHashBuilder.

Examples

use array_map::set::{ArraySet, DefaultHashBuilder, Set};

let set: ArraySet<usize, 31, DefaultHashBuilder> = ArraySet::new();
assert_eq!(set.is_empty(), true);
assert_eq!(set.capacity(), 31);

impl<T, R: RawTable<(T, ())>, B: BuildHasher> ArraySetFacade<T, R, B>

pub fn capacity(&self) -> usize

Returns the number of elements the set can hold in total. The returned value will be equal to the const generic N.

Examples

use array_map::set::{ArraySet, DefaultHashBuilder};

let set: ArraySet<usize, 31, DefaultHashBuilder> = ArraySet::new();
assert_eq!(set.capacity(), 31);