Struct array_map::ArrayMapFacade

pub struct ArrayMapFacade<K, V, R: RawTable<(K, V)>, B = DefaultHashBuilder> { /* private fields */ }

Implementations

impl<K, V, R: RawTable<(K, V)> + Default> ArrayMapFacade<K, V, R, DefaultHashBuilder>

pub fn new() -> Self

Creates an empty ArrayMapFacade with the DefaultHashBuilder.

Examples

use array_map::{ArrayMap, DefaultHashBuilder};

let mut map: ArrayMap<usize, &str, 31, DefaultHashBuilder> = ArrayMap::new();

impl<K, V, R: RawTable<(K, V)> + Default, B: BuildHasher> ArrayMapFacade<K, V, R, B>

pub fn with_hasher(build_hasher: B) -> Self

Creates an empty ArrayMapFacade with the provided BuildHasher.

Note

This is function is identical to ArrayMap::with_build_hasher, the only difference is the name of the function. This function exists for API compatibility with the standard librarys HashMap.

Examples

use array_map::{ArrayMap, DefaultHashBuilder};

let build_hasher = DefaultHashBuilder::default();
let mut map: ArrayMap<usize, &str, 31, DefaultHashBuilder> =
    ArrayMap::with_hasher(build_hasher);

pub fn with_build_hasher(build_hasher: B) -> Self

Creates an empty ArrayMapFacade with the provided BuildHasher.

Examples

use array_map::{ArrayMap, DefaultHashBuilder};

let build_hasher = DefaultHashBuilder::default();
let mut map: ArrayMap<usize, &str, 31, DefaultHashBuilder> =
    ArrayMap::with_build_hasher(build_hasher);

impl<K, V, R: RawTable<(K, V)>, B: BuildHasher> ArrayMapFacade<K, V, R, B>

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold in total.

The returned value, will be equal to the const generic N.

Examples

use array_map::ArrayMap;

let map: ArrayMap<i32, i32, 2> = ArrayMap::new();
assert_eq!(map.capacity(), 2);

let map: ArrayMap<&str, usize, 1234> = ArrayMap::new();
assert_eq!(map.capacity(), 1234);

pub fn len(&self) -> usize

Returns the number of elements in the map.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<u16, &str, 3> = ArrayMap::new();

assert_eq!(map.len(), 0);
map.try_insert(1, "a");
assert_eq!(map.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<_, _, 3> = ArrayMap::new();

assert_eq!(map.is_empty(), true);

map.try_insert(1, "a")?;
assert_eq!(map.len(), 1);
assert_eq!(map.is_empty(), false);

pub fn build_hasher(&self) -> &B

Returns a reference to the map’s BuildHasher.

Examples

use array_map::ArrayMap;
use std::collections::hash_map::RandomState;

let hasher = RandomState::default();
let map: ArrayMap<i32, i32, 12, _> = ArrayMap::with_hasher(hasher);
let hasher: &RandomState = map.build_hasher();

impl<K, V, R, B> ArrayMapFacade<K, V, R, B> where
K: Eq + Hash,
R: RawTable<(K, V)>,
B: BuildHasher,

pub fn entry(&mut self, key: K) -> Result<Entry<'_, K, V, R, B>, CapacityError>

Gets the given key’s corresponding entry in the map for in-place manipulation.

Examples

use array_map::ArrayMap;

// there are exactly 26 letters in the english alphabet
let mut letters: ArrayMap<char, u32, 26> = ArrayMap::new();

// c = character
for c in "a short treatise on fungi".chars() {
    let counter = letters.entry(c)?.or_insert(0);
    *counter += 1;
}

assert_eq!(letters[&'s'], 2);
assert_eq!(letters[&'t'], 3);
assert_eq!(letters[&'u'], 1);
assert_eq!(letters.get(&'y'), None);

pub fn try_insert(
&mut self,
key: K,
value: V
) -> Result<Option<V>, CapacityError>

Inserts a key-value pair into the map.

If the map did not have this key present, None is returned.

If the map did have this key present, the value is updated, and the old value is returned. The key is not updated, though; this matters for types that can be == without being identical. See the module-level documentation for more.

Examples

use array_map::ArrayMap;

#[derive(Debug, Clone, PartialEq)]
enum Status {
    Occupied,
    Empty,
}

let mut castles: ArrayMap<&str, Status, 59> = ArrayMap::new();

// as you can see the status of the "Shimada Castle" is currently unknown:
assert_eq!(castles.get("Shimada Castle"), None);

let castle_entry = castles.entry("Shimada Castle")?;
assert_eq!(castle_entry.insert(Status::Occupied), None);
assert_eq!(castles["Shimada Castle"], Status::Occupied);

// you can also overwrite existing castles:
castles.try_insert("Anvil Castle", Status::Occupied);

let castle_entry = castles.entry("Anvil Castle")?;
assert_eq!(castle_entry.insert(Status::Empty), Some(Status::Occupied));
assert_eq!(castles["Anvil Castle"], Status::Empty);

Note

This is not the same as HashMap::try_insert, which errors if the value is already present in the map.

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<i32, &str, 13> = ArrayMap::new();
map.try_insert(1, "a")?;

assert_eq!(map.get(&1), Some(&"a"));
assert_eq!(map.get(&2), None);

pub fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut V> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns a mutable reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<i32, &str, 13> = ArrayMap::new();

map.try_insert(1, "a")?;

if let Some(x) = map.get_mut(&1) {
    *x = "b";
}
assert_eq!(map[&1], "b");

pub fn clear(&mut self)

Clears the map, removing all key-value pairs.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<u32, &str, 3> = ArrayMap::new();
map.try_insert(1, "a");
assert_eq!(map.is_empty(), false);

map.clear();
assert_eq!(map.is_empty(), true);

pub fn contains_key<Q: ?Sized>(&self, qkey: &Q) -> bool where
K: Borrow<Q>,
Q: Hash + Eq,

Returns true if the map contains a value for the specified key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<i32, &str, 3> = ArrayMap::new();

map.try_insert(1, "a")?;
assert_eq!(map.contains_key(&1), true);
assert_eq!(map.contains_key(&2), false);

pub fn get_key_value<Q: ?Sized>(&self, key: &Q) -> Option<(&K, &V)> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns the key-value pair corresponding to the supplied key.

The supplied key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<i32, &str, 13> = ArrayMap::new();
map.try_insert(1, "a")?;

assert_eq!(map.get_key_value(&1), Some((&1, &"a")));
assert_eq!(map.get_key_value(&2), None);

pub fn get_key_value_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<(&K, &mut V)> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns the key-value pair corresponding to the supplied key.

The supplied key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<i32, &str, 13> = ArrayMap::new();
map.try_insert(1, "a")?;

let (k, v) = map.get_key_value_mut(&1).unwrap();
assert_eq!(k, &1);
assert_eq!(v, &mut "a");
*v = "b";
assert_eq!(map.get_key_value_mut(&1), Some((&1, &mut "b")));
assert_eq!(map.get_key_value_mut(&2), None);

pub fn remove_entry<Q: ?Sized>(&mut self, qkey: &Q) -> Option<(K, V)> where
K: Borrow<Q>,
Q: Hash + Eq,

Removes a key from the map, returning the stored key and value if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<i32, &str, 3> = ArrayMap::new();

map.try_insert(1, "a")?;
assert_eq!(map.remove_entry(&1), Some((1, "a")));
assert_eq!(map.remove_entry(&1), None);

pub fn remove<Q: ?Sized>(&mut self, qkey: &Q) -> Option<V> where
K: Borrow<Q>,
Q: Hash + Eq,

Removes a key from the map, returning the value at the key if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<i32, &str, 3> = ArrayMap::new();

map.try_insert(1, "a")?;
assert_eq!(map.remove(&1), Some("a"));
assert_eq!(map.remove(&1), None);

pub fn get_each_key_value_mut<Q: ?Sized, const M: usize>(
&mut self,
qkeys: [&Q; M]
) -> [Result<(&K, &mut V), UnavailableMutError>; M] where
K: Borrow<Q>,
Q: Hash + Eq,

Attempts to get mutable references to M values in the map at once, with immutable references to the corresponding keys.

Returns an array of length M with the results of each query. For soundness, at most one mutable reference will be returned to any value. An Err(UnavailableMutError::Duplicate(i)) in the returned array indicates that a suitable key-value pair exists, but a mutable reference to the value already occurs at index i in the returned array.

Examples

use array_map::{ArrayMap, UnavailableMutError};

let mut libraries: ArrayMap<&str, usize, 19> = ArrayMap::new();

libraries.try_insert("Bodleian Library", 1602)?;
libraries.try_insert("Athenæum", 1807)?;
libraries.try_insert("Herzogin-Anna-Amalia-Bibliothek", 1691)?;
libraries.try_insert("Library of Congress", 1800)?;

let got = libraries.get_each_key_value_mut([
    "Bodleian Library",
    "Herzogin-Anna-Amalia-Bibliothek",
    "Herzogin-Anna-Amalia-Bibliothek",
    "Gewandhaus",
]);

assert_eq!(
    got,
    [
        Ok((&"Bodleian Library", &mut 1602)),
        Ok((&"Herzogin-Anna-Amalia-Bibliothek", &mut 1691)),
        Err(UnavailableMutError::Duplicate(1)),
        Err(UnavailableMutError::Absent),
    ]
);

pub fn get_each_value_mut<Q: ?Sized, const M: usize>(
&mut self,
qkeys: [&Q; M]
) -> [Result<&mut V, UnavailableMutError>; M] where
K: Borrow<Q>,
Q: Hash + Eq,

Attempts to get mutable references to N values in the map at once.

Returns an array of length N with the results of each query. For soundness, at most one mutable reference will be returned to any value. An Err(UnavailableMutError::Duplicate(i)) in the returned array indicates that a suitable key-value pair exists, but a mutable reference to the value already occurs at index i in the returned array.

Examples

use array_map::{ArrayMap, UnavailableMutError};

let mut libraries: ArrayMap<&str, usize, 19> = ArrayMap::new();
libraries.try_insert("Bodleian Library", 1602)?;
libraries.try_insert("Athenæum", 1807)?;
libraries.try_insert("Herzogin-Anna-Amalia-Bibliothek", 1691)?;
libraries.try_insert("Library of Congress", 1800)?;

let got = libraries.get_each_value_mut([
    "Athenæum",
    "New York Public Library",
    "Athenæum",
    "Library of Congress",
]);
assert_eq!(
    got,
    [
        Ok(&mut 1807),
        Err(UnavailableMutError::Absent),
        Err(UnavailableMutError::Duplicate(0)),
        Ok(&mut 1800),
    ]
);

pub fn drain_filter<F>(&mut self, f: F) -> DrainFilter<'_, K, V, F, R, B> where
F: FnMut(&K, &mut V) -> bool,

Creates an iterator which uses a closure to determine if an element should be removed.

If the closure returns true, the element is removed from the map and yielded. If the closure returns false, or panics, the element remains in the map and will not be yielded.

Note that drain_filter lets you mutate every value in the filter closure, regardless of whether you choose to keep or remove it.

If the iterator is only partially consumed or not consumed at all, each of the remaining elements will still be subjected to the closure and removed and dropped if it returns true.

It is unspecified how many more elements will be subjected to the closure if a panic occurs in the closure, or a panic occurs while dropping an element, or if the DrainFilter value is leaked.

Examples

use array_map::ext::IteratorExt;
use array_map::{array_map, ArrayMap};

let mut map = array_map! {
    "hello" => "hallo",
    "world" => "welt",
    "apple" => "apfel",
    "rust" => "rost",
};

let mut drained: [Option<(_, _)>; 4] =
    map.drain_filter(|k, v| k.len() < 5).try_collect().unwrap();
drained.sort_unstable();

assert_eq!(drained, [None, None, None, Some(("rust", "rost")),]);

pub fn drain(&mut self) -> Drain<'_, K, V, R, B>

Clears the map, returning all key-value pairs as an iterator.

Examples

use array_map::ext::IteratorExt;
use array_map::{array_map, ArrayMap};

let mut map = array_map! {
    "hello" => "hallo",
    "world" => "welt",
    "apple" => "apfel",
};

let mut drained: [Option<(_, _)>; 3] = map.drain().try_collect().unwrap();
drained.sort_unstable();

assert_eq!(
    drained,
    [
        Some(("apple", "apfel")),
        Some(("hello", "hallo")),
        Some(("world", "welt")),
    ]
);

pub fn try_rescale<T, const M: usize>(
self
) -> Result<ArrayMapFacade<K, V, T, B>, RescaleError<M>> where
T: FixedSizeTable<(K, V), M> + Default,

Tries to convert the map with capacity N into a map with capacity M.

Errors

An error will be returned, if the length of the map (ArrayMap::len) is larger than M.

Examples

use array_map::ArrayMap;

let mut map: ArrayMap<char, u32, 1> = ArrayMap::new();

map.try_insert('a', 'a' as u32)?;

// no more values can be inserted

let rescaled: ArrayMap<_, _, 3> = map.try_rescale().expect("failed to rescale");

assert_eq!(rescaled.len(), 1);
assert_eq!(rescaled.capacity(), 3);
assert_eq!(rescaled.get(&'a'), Some(&('a' as u32)));

pub fn retain<F>(&mut self, f: F) where
F: FnMut(&K, &mut V) -> bool,

Retains only the elements specified by the predicate.

Calls the provided function on each entry, removing all entries, where the function returns false.

Examples

use array_map::ArrayMap;

// maps the chars 'a' to 'z' and '0' to '9' to their numerical value
let mut map: ArrayMap<char, u32, { 26 + 10 }> = ArrayMap::new();

for c in 'a'..='z' {
    map.try_insert(c, c as u32)?;
}

for c in '0'..='9' {
    map.try_insert(c, c as u32)?;
}

// only keep chars that are not digits:
map.retain(|c, _| !c.is_ascii_digit());

for c in 'a'..='z' {
    assert_eq!(map.get(&c), Some(&(c as u32)));
}

for c in '0'..='9' {
    assert_eq!(map.get(&c), None);
}

pub fn raw_entry(&self) -> RawEntryBuilder<'_, K, V, R, B>

Creates a raw immutable entry builder for the ArrayMap.

Raw entries provide the lowest level of control for searching and manipulating a map.

They must be manually initialized with a hash and then manually searched.

This is useful for

Hash memoization

Using a search key that does not work with the Borrow trait

Using custom comparison logic with newtype wrappers

Unless you are in such a situation, higher-level and more foolproof APIs like ArrayMap::get should be preferred.

Immutable raw entries have a very limited use; you might instead want to use ArrayMap::raw_entry_mut.

impl<K, V, R, B> ArrayMapFacade<K, V, R, B> where
R: RawTableIter<(K, V)>,
B: BuildHasher,

pub fn iter(&self) -> Iter<'_, K, V, R>

Returns an iterator iterating over the immutable entries of the map.

Examples

use array_map::ext::IteratorExt;
use array_map::{array_map, ArrayMap};

let mut map: ArrayMap<&str, &str, 3> = array_map! {
    "hello" => "你好",
    "good night" => "晚安",
    "good bye" => "再見",
};

let mut iterated: [(&&str, &&str); 3] = map.iter().try_collect()?;
iterated.sort_unstable();

assert_eq!(
    iterated,
    [
        (&"good bye", &"再見"),
        (&"good night", &"晚安"),
        (&"hello", &"你好"),
    ]
);

pub fn iter_mut(&mut self) -> IterMut<'_, K, V, R>

Returns an iterator iterating over the mutable entries of the map.

Examples

use array_map::{array_map, ArrayMap};

let mut map = array_map! {
   0 => 1,
   1 => 2,
   2 => 3,
};

for (key, value) in map.iter_mut() {
    if key % 2 == 0 {
        *value *= 2;
    } else {
        *value += 5;
    }
}

assert_eq!(
    map,
    array_map! {
       0 => 2,
       1 => 7,
       2 => 6,
    }
);

pub fn keys(&self) -> Keys<'_, K, V, R>

An iterator visiting all keys in arbitrary order. The iterator element type is &'a K.

Examples

use array_map::ext::IteratorExt;
use array_map::{array_map, ArrayMap};

let mut map: ArrayMap<&str, &str, 3> = array_map! {
    "hello" => "salut",
    "good night" => "bonne nuit",
    "good bye" => "au revoir",
};

let mut keys: [&&str; 3] = map.keys().try_collect()?;
keys.sort_unstable();

assert_eq!(keys, [&"good bye", &"good night", &"hello",]);

pub fn values(&self) -> Values<'_, K, V, R>

An iterator visiting all values in arbitrary order. The iterator element type is &'a V.

Examples

use array_map::ext::IteratorExt;
use array_map::{array_map, ArrayMap};

let mut map: ArrayMap<&str, &str, 3> = array_map! {
    "hello" => "salut",
    "good night" => "bonne nuit",
    "good bye" => "au revoir",
};

let mut values: [&&str; 3] = map.values().try_collect()?;
values.sort_unstable();

assert_eq!(values, [&"au revoir", &"bonne nuit", &"salut",]);

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V, R>

An iterator visiting all values in arbitrary order. The iterator element type is &mut V.

Examples

use array_map::ext::IteratorExt;
use array_map::{array_map, ArrayMap};

let mut map = array_map! {
   0 => 1,
   1 => 2,
   2 => 3,
};

for value in map.values_mut() {
    if *value % 2 == 0 {
        *value += 1;
    }
}

assert_eq!(
    map,
    array_map! {
       0 => 1,
       1 => 3,
       2 => 3,
    }
);

impl<K, V, const N: usize, B: BuildHasher> ArrayMapFacade<K, V, IndexTable<(K, V), ArrayTable<TableIndex<N>, N>, N>, B> where
K: Hash + Eq,

pub fn truncate(&mut self, n: usize)

Shortens the map, keeping the first n elements and removing the rest.

If n is greater than self.len this does nothing.

Examples

use array_map::IndexMap;

let mut map: IndexMap<&str, &str, 11> = IndexMap::new();

map.try_insert("a", "b")?;
map.try_insert("c", "d")?;
map.try_insert("e", "f")?;

map.truncate(1);
assert_eq!(map.pop(), Some(("a", "b")));
assert_eq!(map.pop(), None);

pub fn shift_remove_entry<Q: ?Sized>(&mut self, qkey: &Q) -> Option<(K, V)> where
K: Borrow<Q>,
Q: Hash + Eq,

Removes the entry from the map and shifts all entries inserted after it to the left.

Examples

use array_map::ext::IteratorExt;
use array_map::IndexMap;

let mut map: IndexMap<i32, &str, 7> = IndexMap::new();

map.try_insert(1, "a")?;
map.try_insert(2, "b")?;
map.try_insert(3, "c")?;
map.try_insert(4, "d")?;

assert_eq!(map.shift_remove_entry(&1), Some((1, "a")));
assert_eq!(map.shift_remove_entry(&1), None);
assert_eq!(
    Ok([(2, "b"), (3, "c"), (4, "d")]),
    map.into_iter().try_collect()
);

Complexity

O(n)

pub fn shift_remove_index(&mut self, index: usize) -> Option<(K, V)>

Removes the entry at the index and shifts down all entries after it.

This preserves the order of the entries.

None is returned if the index is larger than the map.

Examples

use array_map::ext::IteratorExt;
use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.shift_remove_index(0), Some(("apple", "apfel")));

assert_eq!(
    map.iter().try_collect::<[Option<_>; 3]>(),
    Ok([
        Some((&"tree", &"baum")),
        Some((&"cake", &"kuchen")),
        Some((&"food", &"essen")),
    ])
);

assert_eq!(map.shift_remove_index(5), None);

Complexity

O(n)

pub fn pop(&mut self) -> Option<(K, V)>

Removes the last element from the map and returns it.

Examples

use array_map::IndexMap;

let mut map: IndexMap<&str, &str, 11> = IndexMap::new();

map.try_insert("a", "b")?;
map.try_insert("c", "d")?;
map.try_insert("e", "f")?;

assert_eq!(map.pop(), Some(("e", "f")));
assert_eq!(map.pop(), Some(("c", "d")));
assert_eq!(map.pop(), Some(("a", "b")));
assert_eq!(map.pop(), None);

pub fn get_entry_at(&self, index: usize) -> Option<(&K, &V)>

Returns the entry at the index. If the index is larger than or equal to the map’s length, None is returned.

Examples

use array_map::IndexMap;

let mut map: IndexMap<usize, &str, 11> = IndexMap::new();
map.try_insert(0, "a")?;
map.try_insert(1, "b")?;
map.try_insert(2, "c")?;
map.try_insert(3, "d")?;

assert_eq!(map.get_entry_at(0), Some((&0, &"a")));
assert_eq!(map.get_entry_at(3), Some((&3, &"d")));
assert_eq!(map.get_entry_at(4), None);

map.pop();

assert_eq!(map.get_entry_at(3), None);
assert_eq!(map.get_entry_at(2), Some((&2, &"c")));

pub fn get_entry_at_mut(&mut self, index: usize) -> Option<(&K, &mut V)>

Returns the entry at the index. If the index is larger than or equal to the map’s length, None is returned.

Examples

use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

if let Some(entry) = map.get_entry_at_mut(1) {
    assert_eq!(entry.0, &"tree");
    assert_eq!(entry.1, &mut "baum");
    *entry.1 = "bäumchen";
}

assert_eq!(map.get_entry_at_mut(1), Some((&"tree", &mut "bäumchen")));

assert_eq!(map.get_entry_at_mut(5), None);

pub fn swap_indices(&mut self, a: usize, b: usize)

Swaps the position of the two entries a and b.

Panics

If either a or b is out of bounds.

Examples

use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.get_entry_at(1), Some((&"tree", &"baum")));
assert_eq!(map.get_entry_at(2), Some((&"cake", &"kuchen")));

map.swap_indices(1, 2);

assert_eq!(map.get_entry_at(1), Some((&"cake", &"kuchen")));
assert_eq!(map.get_entry_at(2), Some((&"tree", &"baum")));

pub fn try_swap_indices(
&mut self,
a: usize,
b: usize
) -> Result<(), IndexOutOfBoundsError>

Swaps the position of the two entries a and b.

Errors

If either a or b is out of bounds.

Examples

use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.get_entry_at(1), Some((&"tree", &"baum")));
assert_eq!(map.get_entry_at(2), Some((&"cake", &"kuchen")));

assert_eq!(map.try_swap_indices(1, 2), Ok(()));

assert_eq!(map.get_entry_at(1), Some((&"cake", &"kuchen")));
assert_eq!(map.get_entry_at(2), Some((&"tree", &"baum")));

pub fn swap_remove_index(&mut self, index: usize) -> Option<(K, V)>

Removes the entry at the index or returns None if the index is out of bounds by swapping it’s position with the last element and then removing it.

Examples

use array_map::ext::IteratorExt;
use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.swap_remove_index(1), Some(("tree", "baum")));

assert_eq!(
    map.iter().try_collect::<[Option<_>; 3]>(),
    Ok([
        Some((&"apple", &"apfel")),
        Some((&"food", &"essen")),
        Some((&"cake", &"kuchen")),
    ])
);

Complexity

O(1)

pub fn drain_range<R: RangeBounds<usize>>(
&mut self,
range: R
) -> DrainRange<'_, K, V, B, N>

Clears the map in the given index range, returning all entries in that range as an iterator.

This shifts down all the entries following the drained range to fill the gap.

Panics

If the starting point is greater than the end point of if the end point is greater than the length of the map.

Examples

use array_map::ext::IteratorExt;
use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(
    map.drain_range(1..3).try_collect::<[Option<_>; 4]>(),
    Ok([
        //
        Some(("tree", "baum")),
        Some(("cake", "kuchen")),
        None,
        None
    ])
);

assert_eq!(
    map.iter().try_collect::<[Option<_>; 4]>(),
    Ok([
        Some((&"apple", &"apfel")),
        Some((&"food", &"essen")),
        None,
        None
    ])
);

Complexity

O(n^2)