Type Definition array_map::IndexMap

pub type IndexMap<K, V, const N: usize, B = DefaultHashBuilder> = ArrayMapFacade<K, V, IndexTable<(K, V), ArrayTable<TableIndex<N>, N>, N>, B>;

Implementations

impl<K, V, const N: usize, B: BuildHasher> IndexMap<K, V, N, B> where
K: Hash + Eq,

pub fn truncate(&mut self, n: usize)

Shortens the map, keeping the first n elements and removing the rest.

If n is greater than self.len this does nothing.

Examples

use array_map::IndexMap;

let mut map: IndexMap<&str, &str, 11> = IndexMap::new();

map.try_insert("a", "b")?;
map.try_insert("c", "d")?;
map.try_insert("e", "f")?;

map.truncate(1);
assert_eq!(map.pop(), Some(("a", "b")));
assert_eq!(map.pop(), None);

pub fn shift_remove_entry<Q: ?Sized>(&mut self, qkey: &Q) -> Option<(K, V)> where
K: Borrow<Q>,
Q: Hash + Eq,

Removes the entry from the map and shifts all entries inserted after it to the left.

Examples

use array_map::ext::IteratorExt;
use array_map::IndexMap;

let mut map: IndexMap<i32, &str, 7> = IndexMap::new();

map.try_insert(1, "a")?;
map.try_insert(2, "b")?;
map.try_insert(3, "c")?;
map.try_insert(4, "d")?;

assert_eq!(map.shift_remove_entry(&1), Some((1, "a")));
assert_eq!(map.shift_remove_entry(&1), None);
assert_eq!(
    Ok([(2, "b"), (3, "c"), (4, "d")]),
    map.into_iter().try_collect()
);

Complexity

O(n)

pub fn shift_remove_index(&mut self, index: usize) -> Option<(K, V)>

Removes the entry at the index and shifts down all entries after it.

This preserves the order of the entries.

None is returned if the index is larger than the map.

Examples

use array_map::ext::IteratorExt;
use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.shift_remove_index(0), Some(("apple", "apfel")));

assert_eq!(
    map.iter().try_collect::<[Option<_>; 3]>(),
    Ok([
        Some((&"tree", &"baum")),
        Some((&"cake", &"kuchen")),
        Some((&"food", &"essen")),
    ])
);

assert_eq!(map.shift_remove_index(5), None);

Complexity

O(n)

pub fn pop(&mut self) -> Option<(K, V)>

Removes the last element from the map and returns it.

Examples

use array_map::IndexMap;

let mut map: IndexMap<&str, &str, 11> = IndexMap::new();

map.try_insert("a", "b")?;
map.try_insert("c", "d")?;
map.try_insert("e", "f")?;

assert_eq!(map.pop(), Some(("e", "f")));
assert_eq!(map.pop(), Some(("c", "d")));
assert_eq!(map.pop(), Some(("a", "b")));
assert_eq!(map.pop(), None);

pub fn get_entry_at(&self, index: usize) -> Option<(&K, &V)>

Returns the entry at the index. If the index is larger than or equal to the map’s length, None is returned.

Examples

use array_map::IndexMap;

let mut map: IndexMap<usize, &str, 11> = IndexMap::new();
map.try_insert(0, "a")?;
map.try_insert(1, "b")?;
map.try_insert(2, "c")?;
map.try_insert(3, "d")?;

assert_eq!(map.get_entry_at(0), Some((&0, &"a")));
assert_eq!(map.get_entry_at(3), Some((&3, &"d")));
assert_eq!(map.get_entry_at(4), None);

map.pop();

assert_eq!(map.get_entry_at(3), None);
assert_eq!(map.get_entry_at(2), Some((&2, &"c")));

pub fn get_entry_at_mut(&mut self, index: usize) -> Option<(&K, &mut V)>

Returns the entry at the index. If the index is larger than or equal to the map’s length, None is returned.

Examples

use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

if let Some(entry) = map.get_entry_at_mut(1) {
    assert_eq!(entry.0, &"tree");
    assert_eq!(entry.1, &mut "baum");
    *entry.1 = "bäumchen";
}

assert_eq!(map.get_entry_at_mut(1), Some((&"tree", &mut "bäumchen")));

assert_eq!(map.get_entry_at_mut(5), None);

pub fn swap_indices(&mut self, a: usize, b: usize)

Swaps the position of the two entries a and b.

Panics

If either a or b is out of bounds.

Examples

use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.get_entry_at(1), Some((&"tree", &"baum")));
assert_eq!(map.get_entry_at(2), Some((&"cake", &"kuchen")));

map.swap_indices(1, 2);

assert_eq!(map.get_entry_at(1), Some((&"cake", &"kuchen")));
assert_eq!(map.get_entry_at(2), Some((&"tree", &"baum")));

pub fn try_swap_indices(
&mut self,
a: usize,
b: usize
) -> Result<(), IndexOutOfBoundsError>

Swaps the position of the two entries a and b.

Errors

If either a or b is out of bounds.

Examples

use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.get_entry_at(1), Some((&"tree", &"baum")));
assert_eq!(map.get_entry_at(2), Some((&"cake", &"kuchen")));

assert_eq!(map.try_swap_indices(1, 2), Ok(()));

assert_eq!(map.get_entry_at(1), Some((&"cake", &"kuchen")));
assert_eq!(map.get_entry_at(2), Some((&"tree", &"baum")));

pub fn swap_remove_index(&mut self, index: usize) -> Option<(K, V)>

Removes the entry at the index or returns None if the index is out of bounds by swapping it’s position with the last element and then removing it.

Examples

use array_map::ext::IteratorExt;
use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(map.swap_remove_index(1), Some(("tree", "baum")));

assert_eq!(
    map.iter().try_collect::<[Option<_>; 3]>(),
    Ok([
        Some((&"apple", &"apfel")),
        Some((&"food", &"essen")),
        Some((&"cake", &"kuchen")),
    ])
);

Complexity

O(1)

pub fn drain_range<R: RangeBounds<usize>>(
&mut self,
range: R
) -> DrainRange<'_, K, V, B, N>

Clears the map in the given index range, returning all entries in that range as an iterator.

This shifts down all the entries following the drained range to fill the gap.

Panics

If the starting point is greater than the end point of if the end point is greater than the length of the map.

Examples

use array_map::ext::IteratorExt;
use array_map::{index_map, IndexMap};

let mut map: IndexMap<&str, &str, 11> = index_map! {
    @infer,
    "apple" => "apfel",
    "tree" => "baum",
    "cake" => "kuchen",
    "food" => "essen",
}?;

assert_eq!(
    map.drain_range(1..3).try_collect::<[Option<_>; 4]>(),
    Ok([
        //
        Some(("tree", "baum")),
        Some(("cake", "kuchen")),
        None,
        None
    ])
);

assert_eq!(
    map.iter().try_collect::<[Option<_>; 4]>(),
    Ok([
        Some((&"apple", &"apfel")),
        Some((&"food", &"essen")),
        None,
        None
    ])
);

Complexity

O(n^2)