Trait array_map::set::Set

pub trait Set<T> {
    type Error;
    fn len(&self) -> usize;

    fn get<Q: ?Sized>(&self, value: &Q) -> Option<&T>
    where
        T: Borrow<Q>,
        Q: Hash + Eq;

    fn try_replace(&mut self, value: T) -> Result<Option<T>, Self::Error>;

    fn take<Q: ?Sized>(&mut self, value: &Q) -> Option<T>
    where
        T: Borrow<Q>,
        Q: Hash + Eq;

    fn try_insert(&mut self, value: T) -> Result<bool, Self::Error>
    where
        T: Hash + Eq,
    { ... }

    fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool
    where
        T: Borrow<Q>,
        Q: Hash + Eq,
    { ... }

    fn contains<Q: ?Sized>(&self, value: &Q) -> bool
    where
        T: Borrow<Q>,
        Q: Hash + Eq,
    { ... }

    fn is_empty(&self) -> bool { ... }
}

A trait that is implemented by structs that are sets.

So one can for example check if a HashSet and an ArraySet are disjoint.

Associated Types

type Error

Required methods

fn len(&self) -> usize

Returns the number of elements in the set.

fn get<Q: ?Sized>(&self, value: &Q) -> Option<&T> where
T: Borrow<Q>,
Q: Hash + Eq,

Returns a reference to the value in the set, if any, that is equal to the given value.

The value may be any borrowed form of the set’s value type, but Hash and Eq on the borrowed form must match those for the value type.

fn try_replace(&mut self, value: T) -> Result<Option<T>, Self::Error>

Adds a value to the set, replacing the existing value, if any, that is equal to the given one. Returns the replaced value.

Errors

If the set is full or it failed to insert the value.

fn take<Q: ?Sized>(&mut self, value: &Q) -> Option<T> where
T: Borrow<Q>,
Q: Hash + Eq,

Removes and returns the value in the set, if any, that is equal to the given one.

The value may be any borrowed form of the set’s value type, but Hash and Eq on the borrowed form must match those for the value type.

Provided methods

fn try_insert(&mut self, value: T) -> Result<bool, Self::Error> where
T: Hash + Eq,

Adds the value to the set.

If the set did not have this value present, true is returned.

If the set did have this value present, false is returned.

Errors

If the set is full or it failed to insert the value.

fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Hash + Eq,

Removes a value from the set.

Returns true if the value was present in the set and false if it was not in the set.

fn contains<Q: ?Sized>(&self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Hash + Eq,

Returns true if the value is in the set, false if it is not.

fn is_empty(&self) -> bool

Returns true if the set is empty, false if it is not empty.