Trait array_map::set::SetIter

pub trait SetIter<T>: Set<T> + IntoIterator<Item = T> {
    type Iter: Iterator<Item = &'a T>;
    fn iter(&self) -> Self::Iter;

    fn difference<'a, S: Set<T>>(
        &'a self, 
        other: &'a S
    ) -> Difference<'a, T, Self, S>
    where
        Self: Sized,
    { ... }

    fn symmetric_difference<'a, S: SetIter<T>>(
        &'a self, 
        other: &'a S
    ) -> SymmetricDifference<'a, T, Self, S>
    where
        Self: Sized,
        T: Hash + Eq,
    { ... }

    fn intersection<'a, S: Set<T>>(
        &'a self, 
        other: &'a S
    ) -> Intersection<'a, T, Self, S>
    where
        Self: Sized,
        T: Hash + Eq,
    { ... }

    fn union<'a, S: SetIter<T>>(&'a self, other: &'a S) -> Union<'a, T, Self, S>
    where
        Self: Sized,
        T: Hash + Eq,
    { ... }

    fn is_disjoint<S: SetIter<T>>(&self, other: &S) -> bool
    where
        T: Hash + Eq,
    { ... }

    fn is_subset<S: Set<T>>(&self, other: &S) -> bool
    where
        T: Hash + Eq,
    { ... }

    fn is_superset<S: SetIter<T>>(&self, other: &S) -> bool
    where
        T: Hash + Eq,
        Self: Sized,
    { ... }
}

Associated Types

type Iter: Iterator<Item = &'a T>

An immutable iterator over the elements of the set.

Required methods

fn iter(&self) -> Self::Iter

An iterator visiting all elements in arbitrary order. The iterator element type is &'a T.

Provided methods

fn difference<'a, S: Set<T>>(
&'a self,
other: &'a S
) -> Difference<'a, T, Self, S> where
Self: Sized,

Visits the values representing the difference, i.e., the values that are in self but not in other.

In mathematical notation this is self \ other or { x | x ∈ self ∧ x ∉ other }

fn symmetric_difference<'a, S: SetIter<T>>(
&'a self,
other: &'a S
) -> SymmetricDifference<'a, T, Self, S> where
Self: Sized,
T: Hash + Eq,

Visits the values representing the symmetric difference, i.e., the values that are in self or in other but not in both.

In mathematical notation this is self ∪ other \ self ∩ other

fn intersection<'a, S: Set<T>>(
&'a self,
other: &'a S
) -> Intersection<'a, T, Self, S> where
Self: Sized,
T: Hash + Eq,

fn union<'a, S: SetIter<T>>(&'a self, other: &'a S) -> Union<'a, T, Self, S> where
Self: Sized,
T: Hash + Eq,

fn is_disjoint<S: SetIter<T>>(&self, other: &S) -> bool where
T: Hash + Eq,

Returns true if self and other have no elements in common.

This is equivalent to checking for an empty intersection.

In mathematical notation this is self ∩ other = ∅ or (∀x ∈ self: x ∉ other) ∧ (∀x ∈ other: x ∉ self)

fn is_subset<S: Set<T>>(&self, other: &S) -> bool where
T: Hash + Eq,

Returns true if the set is a subset of another, i.e., other contains at least all the values in self.

In mathematical notation this is self ⊆ other.

fn is_superset<S: SetIter<T>>(&self, other: &S) -> bool where
T: Hash + Eq,
Self: Sized,

Returns true if the set is a superset of another, i.e., self contains at least all the values in other.

In mathematical notation this is other ⊆ self.