Type Definition array_map::IndexMap

pub type IndexMap<K, V, const N: usize, B = DefaultHashBuilder> = ArrayMapFacade<K, V, IndexTable<(K, V), ArrayTable<TableIndex<N>, N>, N>, B>;

Implementations

impl<K, V, const N: usize, B: BuildHasher> IndexMap<K, V, N, B> where
K: Hash + Eq,

pub fn truncate(&mut self, n: usize)

Shortens the map, keeping the first n elements and removing the rest.

If n is greater than self.len this does nothing.

Examples

use array_map::IndexMap;

let mut map: IndexMap<&str, &str, 11> = IndexMap::new();

map.insert("a", "b")?;
map.insert("c", "d")?;
map.insert("e", "f")?;

map.truncate(1);
assert_eq!(map.pop(), Some(("a", "b")));
assert_eq!(map.pop(), None);

pub fn shift_remove_entry<Q: ?Sized>(&mut self, qkey: &Q) -> Option<(K, V)> where
K: Borrow<Q>,
Q: Hash + Eq,

Removes the entry from the map and shifts all entries inserted after it to the left.

Examples

use array_map::ext::IteratorExt;
use array_map::IndexMap;

let mut map: IndexMap<i32, &str, 7> = IndexMap::new();

map.insert(1, "a")?;
map.insert(2, "b")?;
map.insert(3, "c")?;
map.insert(4, "d")?;

assert_eq!(map.shift_remove_entry(&1), Some((1, "a")));
assert_eq!(map.shift_remove_entry(&1), None);
assert_eq!(
    Ok([(2, "b"), (3, "c"), (4, "d")]),
    map.into_iter().try_collect()
);

Complexity

O(n)

pub fn pop(&mut self) -> Option<(K, V)>

Removes the last element from the map and returns it.

Examples

use array_map::IndexMap;

let mut map: IndexMap<&str, &str, 11> = IndexMap::new();

map.insert("a", "b")?;
map.insert("c", "d")?;
map.insert("e", "f")?;

assert_eq!(map.pop(), Some(("e", "f")));
assert_eq!(map.pop(), Some(("c", "d")));
assert_eq!(map.pop(), Some(("a", "b")));
assert_eq!(map.pop(), None);

pub fn get_index_entry(&self, index: usize) -> Option<(&K, &V)>

Returns the entry at the index. If the index is larger than or equal to the map’s length, None is returned.

Examples

use array_map::IndexMap;

let mut map: IndexMap<usize, &str, 11> = IndexMap::new();
map.insert(0, "a")?;
map.insert(1, "b")?;
map.insert(2, "c")?;
map.insert(3, "d")?;

assert_eq!(map.get_index_entry(0), Some((&0, &"a")));
assert_eq!(map.get_index_entry(3), Some((&3, &"d")));
assert_eq!(map.get_index_entry(4), None);

map.pop();

assert_eq!(map.get_index_entry(3), None);
assert_eq!(map.get_index_entry(2), Some((&2, &"c")));

pub fn get_index_entry_mut(&mut self, index: usize) -> Option<(&K, &mut V)>